oci: update to spec version 1.0.0-rc6
OCI image spec 1.0.0-rc5 changed the way refs are handled. Instead of
having a refs/ folder containing all refs, there must be a `index.json`
file at the root containing an index of the manifests.

Refs are now encoded using an annotation on the manifest in the
`index.json` file. When creating an emtpy image, we pick 'latest' as the
default ref but this can still be changed using set-tag. If the ref is
set to an empty string, no ref annotation is included in the index.