begin: make local directory OCI begin work
The OCI build mode from a local directory was broken, as files were
simply copied to a rootfs folder under the base image(as in ACI).

Here we instead create an empty OCI image, and perform a copy-to-dir.
While this might be a tiny bit less efficient, we get to reuse the logic
already present in copy-to-dir.