oci: track layer dirtiness before rehashing
Commit d5512af3 introduced lazy rehashing, but always rehashes before a
new layer or at write. Obviously this may introduce an unneeded rehash
when nothing has been changed.

Here we introduce dirtiness tracking when performing copy and run. These
now create a .dirty sibling to the blob folder, in the ociblobs
directory.  When performing the lazy rehash we first check if this file
exists, and skip the rehash if it does not.